Split CompileDesignProcesses into phase-owned files
- Extract connection_lowering, behavioral_trigger_contracts
- Extract spec_planning with self-contained SpecPlan
- Extract spec_session with CompileSpecializations
- Extract layout_pipeline with TopologyPlan
- Extract runtime_data_extraction with products model
- lower.cpp reduced from 2855 to 506 lines
